requirements.txt
# Big Data Processing
pyspark==3.4.1
dask[complete]==2023.5.0
distributed==2023.5.0

# Data Lake
delta-spark==2.4.0

# Data Manipulation
pandas==2.3.3
numpy==1.26.4

# Storage
boto3==1.28.25
minio==7.1.15

# Data Generation
faker==19.2.0

# ML & Tracking
mlflow==2.9.2
scikit-learn==1.3.2

# Formats
fastavro==1.9.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Jupyter
jupyter==1.0.0
notebook==7.0.6
ipywidgets==8.1.1

# Utilities
tqdm==4.66.1
python-dotenv==1.0.0

# Testing
pytest==7.4.3
